+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project aspires to adhere to [Semantic Versioning](https://semver.org/s
 - kokkos@4.7.00
 - mfem@4.8
 
+### Added
+- Added ability to define cameras for rendering using visit camera view parameters
 
 ## [0.9.5] - Released 2025-09-10
 ### Preferred dependency versions for ascent@0.9.5